flask-ai-api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
openai==1.3.0
supabase==2.3.4
pandas==2.1.3
pillow==10.1.0
python-dotenv==1.0.0